backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-cors-headers==4.3.1
django-import-export==3.3.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
Flask==3.0.2
html5lib==1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupPy==1.14
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.4
Pillow==10.1.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
pyHanko==0.22.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMsgBox==1.0.9
pypdf==4.1.0
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
python-bidi==0.4.2
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2023.3.post1
pywhatkit==5.4
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
wikipedia==1.4.0
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
